PEPPER-1642 initial sketch of legacy kit integration with juniper.
